Updated GUI and added Game class
The Game class will serve as my overall 'brain' of the game and handle
pretty much everything moving forward. The Board class will server as
mainly the GUI and I need to reformat and rename everything to make it
cleaner. Will do this next probably because I also need to add each
component in the Board class as a member of the board class so the Game
class can access it.