Refactoring done for the most part
Back from my nap and it was splendid. Most of the refactoring is done,
when I figure out the rest of the game concepts I can fill in the rest.
IE things like the zone names will remain as button1-6 because I do not
know the zone names yet. That will be a small fix once I get there. The
game class is good to go on the main loop. The constructor now
initializes a few things and it will be easy to add on whatever I need
to. Next will be the main game loop I will have to implement some
blocking queues to pass things to the thread.